jwt_compact::alg

Struct Es256k

pub struct Es256k<D = Sha256> { /* private fields */ }

Available on crate feature es256k only.

Algorithm implementing elliptic curve digital signatures (ECDSA) on the secp256k1 curve.

The algorithm does not fix the choice of the message digest algorithm; instead, it is provided as a type parameter. SHA-256 is the default parameter value, but it can be set to any cryptographically secure hash function with 32-byte output (e.g., SHA3-256).

Implementations

impl<D> Es256k<D>
where D: FixedOutputReset<OutputSize = U32> + BlockSizeUser + Clone + Default + HashMarker,

pub fn new(context: Secp256k1<All>) -> Self

Creates a new algorithm instance. This is a (moderately) expensive operation, so if necessary, the algorithm should be clone()d rather than created anew.