Struct jwt_compact::alg::RsaPrivateKey

pub struct RsaPrivateKey { /* private fields */ }

Available on crate feature rsa only.

Represents a whole RSA key, public and private parts.

Implementations

impl RsaPrivateKey

pub fn new<R>(rng: &mut R, bit_size: usize) -> Result<RsaPrivateKey, Error>
where R: CryptoRngCore + ?Sized,

Generate a new Rsa key pair of the given bit size using the passed in rng.

pub fn new_with_exp<R>( rng: &mut R, bit_size: usize, exp: &BigUint, ) -> Result<RsaPrivateKey, Error>
where R: CryptoRngCore + ?Sized,

Generate a new RSA key pair of the given bit size and the public exponent using the passed in rng.

Unless you have specific needs, you should use RsaPrivateKey::new instead.

pub fn from_components( n: BigUint, e: BigUint, d: BigUint, primes: Vec<BigUint>, ) -> Result<RsaPrivateKey, Error>

Constructs an RSA key pair from individual components:

n: RSA modulus

e: public exponent (i.e. encrypting exponent)

d: private exponent (i.e. decrypting exponent)

primes: prime factors of n: typically two primes p and q. More than two primes can be provided for multiprime RSA, however this is generally not recommended. If no primes are provided, a prime factor recovery algorithm will be employed to attempt to recover the factors (as described in NIST SP 800-56B Revision 2 Appendix C.2). This algorithm only works if there are just two prime factors p and q (as opposed to multiprime), and e is between 2^16 and 2^256.

pub fn from_p_q( p: BigUint, q: BigUint, public_exponent: BigUint, ) -> Result<RsaPrivateKey, Error>

Constructs an RSA key pair from its two primes p and q.

This will rebuild the private exponent and the modulus.

Private exponent will be rebuilt using the method defined in NIST 800-56B Section 6.2.1.

pub fn from_primes( primes: Vec<BigUint>, public_exponent: BigUint, ) -> Result<RsaPrivateKey, Error>

Constructs an RSA key pair from its primes.

This will rebuild the private exponent and the modulus.

pub fn to_public_key(&self) -> RsaPublicKey

Get the public key from the private key, cloning n and e.

Generally this is not needed since RsaPrivateKey implements the PublicKey trait, but it can occasionally be useful to discard the private information entirely.

pub fn precompute(&mut self) -> Result<(), Error>

Performs some calculations to speed up private key operations.

pub fn clear_precomputed(&mut self)

Clears precomputed values by setting to None

pub fn crt_coefficient(&self) -> Option<BigUint>

Compute CRT coefficient: (1/q) mod p.

pub fn validate(&self) -> Result<(), Error>

Performs basic sanity checks on the key. Returns Ok(()) if everything is good, otherwise an appropriate error.

pub fn decrypt<P>( &self, padding: P, ciphertext: &[u8], ) -> Result<Vec<u8>, Error>
where P: PaddingScheme,

Decrypt the given message.

pub fn decrypt_blinded<R, P>( &self, rng: &mut R, padding: P, ciphertext: &[u8], ) -> Result<Vec<u8>, Error>
where R: CryptoRngCore, P: PaddingScheme,

Decrypt the given message.

Uses rng to blind the decryption process.

pub fn sign<S>(&self, padding: S, digest_in: &[u8]) -> Result<Vec<u8>, Error>
where S: SignatureScheme,

Sign the given digest.

pub fn sign_with_rng<R, S>( &self, rng: &mut R, padding: S, digest_in: &[u8], ) -> Result<Vec<u8>, Error>
where R: CryptoRngCore, S: SignatureScheme,

Sign the given digest using the provided rng, which is used in the following ways depending on the SignatureScheme:

Pkcs1v15Sign padding: uses the RNG to mask the private key operation with random blinding, which helps mitigate sidechannel attacks.

Pss always requires randomness. Use Pss::new for a standard RSASSA-PSS signature, or Pss::new_blinded for RSA-BSSA blind signatures.