jwt_compact::alg

Trait VerifyingKey

pub trait VerifyingKey<T>: Sized
where
    T: Algorithm<VerifyingKey = Self>,
{
    // Required methods
    fn from_slice(raw: &[u8]) -> Result<Self>;
    fn as_bytes(&self) -> Cow<'_, [u8]>;
}

Verifying key for a specific signature cryptosystem. In the case of public-key cryptosystems, this is a public key.

This trait provides a uniform interface for different backends / implementations of the same cryptosystem.

Required Methods

fn from_slice(raw: &[u8]) -> Result<Self>

Creates a key from raw bytes. Returns an error if the bytes do not represent a valid key.

fn as_bytes(&self) -> Cow<'_, [u8]>

Returns the key as raw bytes.

Implementations should return Cow::Borrowed whenever possible (that is, if the bytes are actually stored within the implementing data structure).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.