jwt_compact::alg

Struct RsaPublicKey

pub struct RsaPublicKey { /* private fields */ }

Available on crate feature rsa only.

Represents the public part of an RSA key.

Implementations

impl RsaPublicKey

pub fn encrypt<R, P>( &self, rng: &mut R, padding: P, msg: &[u8], ) -> Result<Vec<u8>, Error>
where R: CryptoRng + ?Sized, P: PaddingScheme,

Encrypt the given message.

pub fn verify<S>( &self, scheme: S, hashed: &[u8], sig: &[u8], ) -> Result<(), Error>
where S: SignatureScheme,

Verify a signed message.

hashed must be the result of hashing the input using the hashing function passed in through hash.

If the message is valid Ok(()) is returned, otherwise an Err indicating failure.

impl RsaPublicKey

pub const MIN_PUB_EXPONENT: u64 = 2u64

Minimum value of the public exponent e.

pub const MAX_PUB_EXPONENT: u64 = 8_589_934_591u64

Maximum value of the public exponent e.

pub const MAX_SIZE: usize = 4_096usize

Maximum size of the modulus n in bits.

pub fn new(n: BoxedUint, e: BoxedUint) -> Result<RsaPublicKey, Error>

Create a new public key from its components.

This function accepts public keys with a modulus size up to 4096-bits, i.e. RsaPublicKey::MAX_SIZE.

pub fn new_with_max_size( n: BoxedUint, e: BoxedUint, max_size: usize, ) -> Result<RsaPublicKey, Error>

Create a new public key from its components.

pub fn new_unchecked(n: BoxedUint, e: BoxedUint) -> RsaPublicKey

Create a new public key, bypassing checks around the modulus and public exponent size.

This method is not recommended, and only intended for unusual use cases. Most applications should use RsaPublicKey::new or RsaPublicKey::new_with_max_size instead.