Trait jwt_compact::AlgorithmSignature

pub trait AlgorithmSignature: Sized {
    fn try_from_slice(slice: &[u8]) -> Result<Self>;

    fn as_bytes(&self) -> Cow<[u8]>;
}

Signature for a certain JWT signing Algorithm.

We require that signature can be restored from a byte slice, and can be represented as a byte slice.

Required methods

fn try_from_slice(slice: &[u8]) -> Result<Self>

Attempts to restore a signature from a byte slice. This method may fail if the slice is malformed (e.g., has a wrong length).

fn as_bytes(&self) -> Cow<[u8]>

Represents this signature as bytes.

Loading content...